[leancode_lint] Make `catch_parameter_names` lint configurable
This change allows for the configuration of the preferred names for exception and stack trace parameters in catch clauses through the `analysis_options.yaml` file.

The `CatchParameterNames` lint now reads its configuration from the custom lint configs, with `err` and `st` remaining the default names for the exception and stack trace respectively.